Uploaded original code
Uploaded the original code from the original day that I started it. I so far only got a method to download the img from the URL and obtain a URL from the user. Need to work on this from a non-work computer since I need admin rights to download a HTML Parser